requirements.txt
pyserial==3.5
pynput==1.7.6
mysql-connector-python==8.0.33
protobuf==3.20.3
PyQt5==5.15.9
paho-mqtt==1.6.1